requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
openai>=1.35.0
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
asyncio==3.4.3
aiohttp==3.9.1
webdriver-manager==4.0.1
retry==0.9.2
tqdm==4.66.1
json5==0.9.14